Temperature reimplemented, will fix combined tomorrow.

+++ INFT3970-Project/INFT3970-Project/Models/Proceedure Models/AverageTemperatureModelWithId.cs
@@ -8,7 +8,7 @@ namespace INFT3970Project.Models
     public class AverageTemperatureModelWithId
     {
         public int SensorId { get; set; }
-        public double Temperature { get; set; }
+        public decimal Temperature { get; set; }
         public DateTime StartTime { get; set; }
         public DateTime EndTime { get; set; }
     }
